changes for cross project use
